[Chore] 컴포넌트명, 함수명 수정
