+++ src/types/search.ts
@@ -1,0 +1,4 @@
+export interface SubwayType {
+  line_num: string | string[];
+  station_nm: string;
+}
